# Optimize usage of `Stack<T>` in SyntaxNavigator