@jaredpar 2024-08-21
> The fixed method is faster. Below are the results of basic benchmarks that follow the above control flows:

Can you post the benchmark that you wrote?