@steveberdy 2024-08-21
@jaredpar sure. It's just a bare-minimum replication of the general control flow
```csharp
public class Benchmarks
{
    private static readonly Stack<string> s = new();

    static Benchmarks()
    {
        for (int i = 0; i < 1000; i++)
        {
            s.Push(Guid.NewGuid().ToString("N"));
        }
    }

    [Benchmark]
    public string Peek_AndPopIfSuccess()
    {
        string t = s.Peek();
        if (t.Contains("11"))
        {
            s.Pop();
            return t;
        }
        return t;
    }

    [Benchmark]
    public string Pop_AndPushIfFail()
    {
        string t = s.Pop();
        if (t.Contains("11"))
        {
            return t;
        }
        s.Push(t);
        return t;
    }
}
```