@jaredpar 2024-08-21
@steveberdy that's a benchmark for `Stack<T>` directly, where is the benchmark about how this change improves the C# code base?