@CyrusNajmabadi 2024-08-22
possibly a stack overflow.

give that this simple test project fails:

![image](https://github.com/user-attachments/assets/61ad91ea-194f-4299-a309-39c2d89348b6)

you can just run those tests in the IDE:

![image](https://github.com/user-attachments/assets/9e47d4d2-306d-4fae-b1b0-b85d0bc7f25a)

and see what test it fails on, and just debug that test.